Adds Action to Select All Files in a Dir
Adds a new method for the SelectDirButton to produce a
folder browser and collect the files from the path provided.
These files are then added to List of filenames. Each use of this
method or the selecting individual filenames clears the list, so
that a fresh start can be obtained. Adds additional error
handling for when a user cancels out of a folder browser and
when creating the new files.